urlprocessor/pyproject.toml
[tool.poetry]
name = "urlprocessor"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.11.9"
kafka-python = "^2.0.2"
langchain = "0.1.14"
langchain-openai = "0.1.1"
beautifulsoup4 = "^4.12.3"
weaviate-client = "4.5.4"
redis = "5.0.3"
six = "^1.16.0"
langchain-anthropic = "^0.1.11"
nest-asyncio = "^1.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
